[stateless_validation] Split stateless validation primitives into module (#11845)
Refactoring only, no functional changes. The stateless validation
primitives file was getting too complex.

This PR splits the structs into separate files and organizes into
module.
