+++ chain/client/src/stateless_validation/chunk_validator/orphan_witness_pool.rs
@@ -3,7 +3,8 @@ use std::num::NonZeroUsize;
 use lru::LruCache;
 use near_chain_configs::default_orphan_state_witness_pool_size;
 use near_primitives::hash::CryptoHash;
-use near_primitives::stateless_validation::{ChunkProductionKey, ChunkStateWitness};
+use near_primitives::stateless_validation::state_witness::ChunkStateWitness;
+use near_primitives::stateless_validation::ChunkProductionKey;
 use near_primitives::types::BlockHeight;
 
 use metrics_tracker::OrphanWitnessMetricsTracker;
@@ -121,7 +122,7 @@ impl Default for OrphanStateWitnessPool {
 }
 
 mod metrics_tracker {
-    use near_primitives::stateless_validation::ChunkStateWitness;
+    use near_primitives::stateless_validation::state_witness::ChunkStateWitness;
 
     use crate::metrics;
 
@@ -178,7 +179,7 @@ mod metrics_tracker {
 mod tests {
     use near_primitives::hash::{hash, CryptoHash};
     use near_primitives::sharding::{ShardChunkHeader, ShardChunkHeaderInner};
-    use near_primitives::stateless_validation::ChunkStateWitness;
+    use near_primitives::stateless_validation::state_witness::ChunkStateWitness;
     use near_primitives::types::{BlockHeight, ShardId};
 
     use super::OrphanStateWitnessPool;
