+++ chain/client/src/stateless_validation/partial_witness/partial_witness_tracker.rs
@@ -8,10 +8,11 @@ use near_chain::chain::ChunkStateWitnessMessage;
 use near_chain::Error;
 use near_epoch_manager::EpochManagerAdapter;
 use near_o11y::log_assert_fail;
-use near_primitives::stateless_validation::{
-    ChunkProductionKey, ChunkStateWitness, ChunkStateWitnessSize, EncodedChunkStateWitness,
-    PartialEncodedStateWitness,
+use near_primitives::stateless_validation::partial_witness::PartialEncodedStateWitness;
+use near_primitives::stateless_validation::state_witness::{
+    ChunkStateWitness, ChunkStateWitnessSize, EncodedChunkStateWitness,
 };
+use near_primitives::stateless_validation::ChunkProductionKey;
 use time::ext::InstantExt as _;
 
 use crate::client_actor::ClientSenderForPartialWitness;
