+++ core/primitives/src/stateless_validation/chunk_endorsement.rs
@@ -1,0 +1,59 @@
+use std::fmt::Debug;
+
+use crate::sharding::ChunkHash;
+use crate::validator_signer::ValidatorSigner;
+use borsh::{BorshDeserialize, BorshSerialize};
+use near_crypto::{PublicKey, Signature};
+use near_primitives_core::types::AccountId;
+
+use super::SignatureDifferentiator;
+
+/// The endorsement of a chunk by a chunk validator. By providing this, a
+/// chunk validator has verified that the chunk state witness is correct.
+#[derive(Debug, Clone, PartialEq, Eq, BorshSerialize, BorshDeserialize)]
+pub struct ChunkEndorsement {
+    inner: ChunkEndorsementInner,
+    pub account_id: AccountId,
+    pub signature: Signature,
+}
+
+impl ChunkEndorsement {
+    pub fn new(chunk_hash: ChunkHash, signer: &ValidatorSigner) -> ChunkEndorsement {
+        let inner = ChunkEndorsementInner::new(chunk_hash);
+        let account_id = signer.validator_id().clone();
+        let signature = signer.sign_chunk_endorsement(&inner);
+        Self { inner, account_id, signature }
+    }
+
+    pub fn verify(&self, public_key: &PublicKey) -> bool {
+        let data = borsh::to_vec(&self.inner).unwrap();
+        self.signature.verify(&data, public_key)
+    }
+
+    pub fn chunk_hash(&self) -> &ChunkHash {
+        &self.inner.chunk_hash
+    }
+
+    pub fn validate_signature(
+        chunk_hash: ChunkHash,
+        signature: &Signature,
+        public_key: &PublicKey,
+    ) -> bool {
+        let inner = ChunkEndorsementInner::new(chunk_hash);
+        let data = borsh::to_vec(&inner).unwrap();
+        signature.verify(&data, public_key)
+    }
+}
+
+/// This is the part of the chunk endorsement that is actually being signed.
+#[derive(Debug, Clone, PartialEq, Eq, BorshSerialize, BorshDeserialize)]
+pub struct ChunkEndorsementInner {
+    chunk_hash: ChunkHash,
+    signature_differentiator: SignatureDifferentiator,
+}
+
+impl ChunkEndorsementInner {
+    fn new(chunk_hash: ChunkHash) -> Self {
+        Self { chunk_hash, signature_differentiator: "ChunkEndorsement".to_owned() }
+    }
+}
